Feature: hashmap and test file
The test file contains some method calls from an
instantiated object in order to make sure the
HashMap class is working as expected. It should
be better to use a proper test file instead of
the .rb one used in this project.

Also updated README.md to remember some ideas
when returning to this project.